+++ src/lib/components/json-schema/form/index.tsx
@@ -4,18 +4,16 @@
 /* eslint-disable no-continue */
 /* eslint-disable no-console */
 /* eslint-disable no-restricted-syntax */
-import {
+import Form, {
   Templates as DefaultTemplates,
   Widgets as DefaultWidgets,
-  Form,
 } from "@rjsf/chakra-ui";
 import type { FormProps } from "@rjsf/core";
 import type { RJSFSchema, RJSFValidationError } from "@rjsf/utils";
-import { createSchemaUtils } from "@rjsf/utils";
 import { customizeValidator } from "@rjsf/validator-ajv8";
-import isEqual from "lodash/isEqual";
+import { isUndefined } from "lodash";
 import type { FC } from "react";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 
 import type { JsonDataType } from "lib/types";
 
@@ -29,66 +27,13 @@ const v8Validator = customizeValidator({
   ajvOptionsOverrides: { validateFormats: false },
 });
 
-function deleteExtraneousOneOfKeys(value: Record<string, unknown>) {
-  const actualOneOfOptions = Object.keys(value);
+function fixSchema(schema: RJSFSchema) {
+  if (schema.type === "object" && isUndefined(schema.properties))
+    schema.properties = {};
 
-  // check if there are TOO many options selected in the value
-  if (actualOneOfOptions.length === 2) {
-    console.log("Deleting", actualOneOfOptions[0]);
-    delete value[actualOneOfOptions[0]];
-  } else if (actualOneOfOptions.length > 2) {
-    console.warn("Unexpected number of oneOf options", actualOneOfOptions);
-    // TODO: throw an error?
-  }
-}
-
-/**
- * A helper method to fix a bug in the rjsf library.
- * The bug is that `oneOf` properties will ALWAYS contain the first option.
- * @param formData
- * @param collapsedSchema - the schema that has been collapsed via `createSchemaUtils`
- */
-function fixOneOfKeys(
-  formData: Record<string, unknown>,
-  collapsedSchema: RJSFSchema
-) {
-  // if the entry is supposed to be a oneof *itself*, then check that it only has one key
-  if (collapsedSchema.oneOf) {
-    deleteExtraneousOneOfKeys(formData);
-
-    // Now recursively check the keys (though there should only be one)
-    for (const [key, value] of Object.entries(formData)) {
-      if (typeof value === "object") {
-        // Find the schema for the key('s value)
-        const valueSchema = collapsedSchema.oneOf
-          .filter((i): i is RJSFSchema => typeof i === "object")
-          .find((i) => i.required?.length === 1 && i.required[0] === key);
-
-        if (!valueSchema) return;
-        fixOneOfKeys(value as Record<string, unknown>, valueSchema);
-      }
-    }
-    return;
-  }
-
-  // iterate through each entry in the formData to check if it's oneOf
-  for (const [key, value] of Object.entries(formData)) {
-    // skip non-objects
-    if (!value || typeof value !== "object") continue;
-
-    const valueSchema = collapsedSchema.properties?.[key];
-    // Skip those without a valid schema
-    if (!valueSchema || typeof valueSchema === "boolean") continue;
-
-    // if the entry is supposed to be a oneof, then check that it only has one key
-    if (valueSchema.oneOf) {
-      // console.log("Found oneOf", key, value);
-      deleteExtraneousOneOfKeys(value as Record<string, unknown>);
-    }
-
-    // Since we know that the value is an object, we can recurse
-    fixOneOfKeys(value as Record<string, unknown>, valueSchema);
-  }
+  Object.values(schema).forEach((value) => {
+    if (typeof value === "object") fixSchema(value);
+  });
 }
 
 export interface JsonSchemaFormProps
@@ -119,37 +64,17 @@ export const JsonSchemaForm: FC<JsonSchemaFormProps> = ({
 }) => {
   const [formData, setFormData] = useState<JsonDataType>(initialFormData);
 
-  const collapsedSchema = useMemo(
-    () => createSchemaUtils(v8Validator, schema).retrieveSchema(schema),
-    [schema]
-  );
-
-  const fixOneOfKeysCallback = useCallback(
-    (data: Record<string, unknown>) => fixOneOfKeys(data, collapsedSchema),
-    [collapsedSchema]
-  );
-
   const onSubmit = (values: Record<string, unknown>) => {
-    fixOneOfKeysCallback(values);
     console.log("onSubmit", values);
-
     propsOnSubmit?.(values);
   };
 
   const onChange = useCallback(
     (data: JsonDataType, errors: RJSFValidationError[]) => {
-      if (data) {
-        const values: JsonDataType = structuredClone(data);
-        if (typeof values === "object")
-          fixOneOfKeysCallback(values as Record<string, unknown>);
-
-        if (!isEqual(formData, values)) {
-          setFormData(values);
-          propsOnChange?.(values, errors);
-        }
-      }
+      setFormData(data);
+      propsOnChange?.(data, errors);
     },
-    [fixOneOfKeysCallback, formData, propsOnChange]
+    [formData, propsOnChange]
   );
 
   useEffect(() => {
@@ -160,6 +85,7 @@ export const JsonSchemaForm: FC<JsonSchemaFormProps> = ({
     propsOnChange?.(initialFormData, errors);
   }, [JSON.stringify(initialFormData)]);
 
+  fixSchema(schema);
   return (
     <Form
       id={formId}
@@ -198,6 +124,10 @@ export const JsonSchemaForm: FC<JsonSchemaFormProps> = ({
         onSubmit(values);
       }}
       onError={() => console.error("errors")}
+      experimental_defaultFormStateBehavior={{
+        // Assign value to formData when only default is set
+        emptyObjectFields: "skipEmptyDefaults",
+      }}
     />
   );
 };