+++ src/lib/components/json-schema/form/templates/TitleFieldTemplate.tsx
@@ -1,13 +1,21 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Text } from "@chakra-ui/react";
-import type { TitleFieldProps } from "@rjsf/utils";
+import type {
+  FormContextType,
+  RJSFSchema,
+  StrictRJSFSchema,
+  TitleFieldProps,
+} from "@rjsf/utils";
+
 /** The `TitleField` is the template to use to render the title of a field
  *
  * @param props - The `TitleFieldProps` for this component
  */
-export default function TitleFieldTemplate<T = any, F = any>(
-  props: TitleFieldProps<T, F>
-) {
+export default function TitleField<
+  T = any,
+  S extends StrictRJSFSchema = RJSFSchema,
+  F extends FormContextType = any,
+>(props: TitleFieldProps<T, S, F>) {
   const { id, title } = props;
   return (
     <Text id={id} variant="body3" fontWeight={700}>